+++ freenom.sh
@@ -11,7 +11,7 @@
 # This is free software, and you are welcome to redistribute it               #
 # under certain conditions.                                                   #
 # See LICENSE file for more information                                       #
-# gpl-3.0-only                                                  v2023-06-10   #
+# gpl-3.0-only                                                  v2023-08-03   #
 ###############################################################################
 
 # shellcheck disable=SC2317
@@ -218,11 +218,11 @@ if [ "${freenom_http_resolve:-0}" -eq 1 ]; then
 fi
 
 # AWS WAF CAPTCHA token. To manually get it from browser:
-# goto my.freenom.com, solve captcha puzzle and copy cookie (valid 3 mins)
+# goto my.freenom.com, solve captcha puzzle and copy cookie 'aws-waf-token' (valid ~3 mins)
 if [ -n "$AWS_WAF_TOKEN" ]; then
-  AWS_WAF_TOKEN="${AWS_WAF_TOKEN#aws-waf-token?} )"
-  if echo "$AWS_WAF_TOKEN" | grep -Eq '[0-9a-f-]{36}:[A-Za-z0-9+/]{16}:[A-Za-z0-9+/]{227}='; then
-    curlExtraOpts+=" -b aws-waf-token=$AWS_WAF_TOKEN "
+  TOKEN="${AWS_WAF_TOKEN#aws-waf-token=}"
+  if echo "$TOKEN" | grep -Eq '[0-9a-f-]{36}:[A-Za-z0-9+/]{16}:[A-Za-z0-9+/=]{64,}'; then
+    curlExtraOpts+=" -b aws-waf-token=${TOKEN} "
   else
     echo "Error: Incorrect AWS WAF CAPTCHA token"
     exit 1